add: add necessary packages, set uint for primarykey

+++ finexblock/entity/instance/finexblock_ip.model.go
@@ -3,8 +3,8 @@ package instance
 import "gorm.io/gorm"
 
 type FinexblockServerIP struct {
-	ID       uint64 `gorm:"column:id;primary_key;auto_increment;comment:'기본키'"`
-	ServerID uint64 `gorm:"column:server_id;not null;comment:'서버 id'"`
+	ID       uint   `gorm:"column:id;primary_key;auto_increment;comment:'기본키'"`
+	ServerID uint   `gorm:"column:server_id;not null;comment:'서버 id'"`
 	IP       string `gorm:"column:ip;type:longtext;not null;comment:'IP 주소'"`
 	gorm.Model
 }
