+++ finexblock/entity/order/snapshot_order_book.go
@@ -5,7 +5,7 @@ import (
 )
 
 type SnapshotOrderBook struct {
-	ID            uint64    `gorm:"primary_key;auto_increment;comment:'기본키'" json:"id"`
+	ID            uint      `gorm:"primary_key;auto_increment;comment:'기본키'" json:"id"`
 	OrderSymbolID uint      `gorm:"not null;comment:'코인쌍';index" json:"order_symbol_id"` // Index added for foreign key
 	BidOrderList  string    `gorm:"type:longtext;not null;comment:'매수주문리스트 문자열'" json:"bid_order_list"`
 	AskOrderList  string    `gorm:"type:longtext;not null;comment:'매도주문리스트 문자열'" json:"ask_order_list"`
